Added constrain tree capacities to Phyml, Raxml and Fasttree tasks. This solves the problem of multi seq outgroups. Added statal functionality for faster alg analysis. Added nprtop utility

+++ nprlib/master_job.py
@@ -157,6 +157,7 @@ class Job(object):
         open(self.status_file, "w").write(status)
         
     def clean(self):
-        shutil.rmtree(self.jobdir)
+        if os.path.exists(self.jobdir):
+            shutil.rmtree(self.jobdir)
